requirements.txt
dj-database-url==0.5.0
Django==2.2
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
